+++ src/fr/zetioz/zefreeze/listeners/ZeFreezeBlockPlaceListener.java
@@ -1,4 +1,4 @@
-package fr.zetioz.zefreeze.listener;
+package fr.zetioz.zefreeze.listeners;
 
 import fr.zetioz.coreutils.FilesManagerUtils;
 import fr.zetioz.zefreeze.ZeFreezePlugin;
@@ -6,19 +6,19 @@ import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.event.block.BlockPlaceEvent;
 
 import java.io.FileNotFoundException;
 
 import static fr.zetioz.coreutils.ColorUtils.sendMessage;
 
-public class ZeFreezePlayerInteractListener implements Listener, FilesManagerUtils.ReloadableFiles {
+public class ZeFreezeBlockPlaceListener implements Listener, FilesManagerUtils.ReloadableFiles {
     private final ZeFreezePlugin instance;
     private YamlConfiguration messages;
     private YamlConfiguration config;
     private String prefix;
 
-    public ZeFreezePlayerInteractListener(final ZeFreezePlugin instance) throws FileNotFoundException {
+    public ZeFreezeBlockPlaceListener(final ZeFreezePlugin instance) throws FileNotFoundException {
         this.instance = instance;
         instance.getFilesManagerUtils().addReloadable(this);
         reloadFiles();
@@ -32,13 +32,13 @@ public class ZeFreezePlayerInteractListener implements Listener, FilesManagerUti
     }
 
     @EventHandler
-    public void onPlayerInteract(PlayerInteractEvent event) {
+    public void onBlockPlace(BlockPlaceEvent event) {
         final Player player = event.getPlayer();
 
         if (!instance.getPlayerFrozen().containsKey(player.getUniqueId())
-                || config.getBoolean("disable-interaction", false)) return;
+                || !config.getBoolean("disable-block-place", false)) return;
 
         event.setCancelled(true);
-        sendMessage(player, messages.getStringList("errors.interact-while-frozen"), prefix);
+        sendMessage(player, messages.getStringList("errors.place-while-frozen"), prefix);
     }
 }
