+++ src/fr/zetioz/zefreeze/listeners/ZeFreezeInventoryClickListener.java
@@ -1,7 +1,7 @@
-package fr.zetioz.zefreeze.listener;
+package fr.zetioz.zefreeze.listeners;
 
 import fr.zetioz.zefreeze.ZeFreezePlugin;
-import fr.zetioz.zefreeze.gui.AntiDisconnectionGUI;
+import fr.zetioz.zefreeze.guis.AntiDisconnectionGUI;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
