+++ src/fr/zetioz/zefreeze/listeners/ZeFreezeInventoryCloseListener.java
@@ -1,8 +1,8 @@
-package fr.zetioz.zefreeze.listener;
+package fr.zetioz.zefreeze.listeners;
 
 import fr.zetioz.coreutils.FilesManagerUtils;
 import fr.zetioz.zefreeze.ZeFreezePlugin;
-import fr.zetioz.zefreeze.gui.AntiDisconnectionGUI;
+import fr.zetioz.zefreeze.guis.AntiDisconnectionGUI;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
